Enable keyboard movement when multiple items are allowed in drag-and-drop
Keyboard navigation (with arrow keys) didn’t work when 'AllowMultipleItemsInSamePosition' was selected. The missing behavior was implemented mainly in the handleKeyDown() function. Now, users can move items up and down using the keyboard, even inside or between groups. If an item reaches the end of a group, it moves out and creates or merges with another group.